Tabs

+++ frameworks/bridge/cj_frontend/interfaces/cj_ffi/cj_tab_ffi.h
@@ -117,7 +117,7 @@ struct CJTabsBarBackgroundBlurStyleOptions {
     int32_t colorMode;
     int32_t adaptiveColor;
     CJTabsBlurOptions blurOptions;
-    double scale;
+    float scale;
     int32_t policy;
     uint32_t inactiveColor;
 };
